services/bu1_onboarding/pyproject.toml
[project]
name = "ritecare-bu1-onboarding"
version = "0.1.0"
description = "RiteCare BU1 - Customer Onboarding microservice"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "motor>=3.5",
    "pymongo>=4.8",
    "pydantic>=2.8",
    "pydantic-settings>=2.4",
    "slowapi>=0.1.9",
    "google-genai>=1.0",
    "pypdf>=4.0",
    "tiktoken>=0.7",
    "python-dotenv>=1.0",
    "structlog>=24.0",
    "pydantic[email]",
    "pypdf"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "ruff>=0.6",
    "mypy>=1.11",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api", "service", "dao", "common"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.pytest.ini_options]
asyncio_mode = "auto"